refactor: Move type definitions to root of @types
