+++ src/syntax/elements/core/elementInstruction.ts
@@ -1,11 +1,7 @@
-import {
-    IElementBlock,
-    IElementInstruction,
-    IElementStatement
-} from '@/@types/syntax/elementsCore';
-import { TData, TDataName } from '@/@types/syntax/data';
+import { TData, TDataName } from '@/@types/data';
 import { ElementSyntax } from './elementSyntax';
-import { TElementName } from '@/@types/syntax/elementSpecification';
+import { TElementName } from '@/@types/elementSpecification';
+import { IElementBlock, IElementInstruction, IElementStatement } from '@/@types/elementsCore';
 
 /**
  * @virtual
