+++ src/syntax/elements/core/elementSyntax.ts
@@ -1,6 +1,6 @@
-import { IElementSyntax } from '@/@types/syntax/elementsCore';
-import { TDataName } from '@/@types/syntax/data';
-import { TElementName } from '@/@types/syntax/elementSpecification';
+import { TDataName } from '@/@types/data';
+import { TElementName } from '@/@types/elementSpecification';
+import { IElementSyntax } from '@/@types/elementsCore';
 
 /**
  * @virtual
