+++ src/syntax/elements/program/values/elementValue.ts
@@ -1,5 +1,5 @@
-import { TDataName } from '@/@types/syntax/data';
-import { TElementName } from '@/@types/syntax/elementSpecification';
+import { TDataName } from '@/@types/data';
+import { TElementName } from '@/@types/elementSpecification';
 import { ElementData } from '../../core/elementArgument';
 
 /**
